backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
pydantic==2.7.0
openai>=1.0.0
apscheduler==3.10.4
python-dotenv==1.0.1
aiosqlite==0.20.0
bcrypt>=4.1.0
python-jose[cryptography]==3.3.0
email-validator>=2.2.0
